pyproject.toml
[project]
name = "kompello"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.5",
    "django-allauth[mfa]>=65.4.1",
    "django-auditlog>=3.0.0",
    "django-cors-headers>=4.7.0",
    "django-filter>=24.3",
    "django-guardian>=2.4.0",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
]
